+++ Cult of the Lamb/MysticAssistantRedux/Plugin.cs
@@ -6,7 +6,7 @@ public class Plugin : BaseUnityPlugin
 {
     private const string PluginGuid = "p1xel8ted.cotl.mysticassistant";
     private const string PluginName = "Mystic Assistant Redux";
-    private const string PluginVer = "0.1.1";
+    private const string PluginVer = "0.1.2";
 
     internal static ManualLogSource Log { get; private set; }